% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allenpvc.R
\name{allenpvc}
\alias{allenpvc}
\title{Adult mouse cortical cell taxonomy by single cell transcriptomics}
\format{The data is encapsulated in a
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object available
through \code{\link[ExperimentHub]{ExperimentHub}}}
\source{
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71585
}
\description{
Celular taxonomy of the primary visual cortex in adult mice based on single
cell RNA-sequencing from a study performed by the Allen Institute for Brain
Science. In said study 49 transcriptomic cell types are identified.
This data set is the supplementary data from GEO accession
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE71585}{GSE71585}
encapsulated in a \code{\link[SingleCellExperiment]{SingleCellExperiment}}.
}
\details{
See the vignette for examples of using these data in differential gene
expression analysis.

browseVignettes("allenpvc")

Details of how this data was created are in the inst/scripts/ directory of
the source package.
}
\examples{
allenpvc()
}
\references{
Tasic, Bosiljka, et al. Adult mouse cortical cell taxonomy
    revealed by single cell transcriptomics. Nature neuroscience 19.2 (2016):
    335.
}
